Fixes navigation issue with about page
Attempted to Create Ative class to show user which page is active. Class is only on index and about pages. The Class does NOT work. There is an issue getting the background to change on active page. Either issue with class or using the wrong type of identifier. Issue needs to be addressed.